Patient Health Questionnaire-4 Added
The 'Patient Health Questionnaire-4' (PHQ-4) is a 4 item inventory rated on a 4 point Likert-type scale. 
Its items are drawn from the first two items of the 'Generalized Anxiety Disorder–7 scale' 
(GAD–7) and the 'Patient Health Questionnaire-8' (PHQ-8). 
Its purpose is to allow for very brief and accurate measurement of depression and anxiety.